requirements.txt
alembic==1.13.2
alembic_utils==0.8.4
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colour==0.1.5
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.112.0
fastf1==3.4.0
filelock==3.16.1
firebase-admin==6.5.0
flupy==1.2.0
fonttools==4.53.1
fpdf2==2.7.9
furl==2.1.3
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gotrue==2.6.2
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.3.0
mailjet-rest==1.3.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
metahopt==1.1.1
msgpack==1.1.0
nest-asyncio==1.6.0
-e git+https://github.com/Nomiks7/nomiks-builder-api@acebaeaa52f43b23d061c5a55057ce878008db33#egg=nomiks_builder_api
numpy==2.0.1
oauthlib==3.2.2
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parse==1.20.2
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
postgrest==0.16.9
prompt_toolkit==3.0.47
Protego==0.3.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.5.3
pydantic-extra-types==2.9.0
pydantic_core==2.14.6
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pypdf==4.3.1
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
queuelib==1.7.0
RapidFuzz==3.10.0
realtime==1.0.6
requests==2.32.3
requests-cache==1.2.1
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.14.0
Scrapy==2.11.2
seaborn==0.13.2
selenium==4.25.0
sentry-sdk==2.12.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
stack-data==0.6.3
starlette==0.37.2
starlette-admin==0.14.1
storage3==0.7.7
StrEnum==0.4.15
stripe==10.7.0
structlog==24.4.0
supabase==2.6.0
supafunc==0.5.1
temp-mail==0.2
tempMail2==1.0.2
timple==0.1.8
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
w3lib==2.2.1
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
zope.interface==7.0.3
